Refactor project-cli init wizard
